+++ build.gradle
@@ -1,43 +1,44 @@
-buildscript {
-	repositories {
-		mavenCentral()
-	}
-}
 plugins {
 	id 'java'
 	id 'org.springframework.boot' version '3.2.3'
 	id 'io.spring.dependency-management' version '1.1.4'
 }
 
+group = 'WebProject'
+version = '0.0.1-SNAPSHOT'
 
-allprojects {
-	group = 'WebProject'
-	version = '0.0.1-SNAPSHOT'
+java {
 	sourceCompatibility = '17'
 }
 
-subprojects {
-	apply plugin: "java"
-	apply plugin: 'java-library'
-	apply plugin: "io.spring.dependency-management"
-	apply plugin: "org.springframework.boot"
-
-	repositories {
-		mavenCentral()
+configurations {
+	compileOnly {
+		extendsFrom annotationProcessor
 	}
+}
 
-	configurations {
-		compileOnly {
-			extendsFrom annotationProcessor
-		}
-	}
-	dependencies {
-		implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-		implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
+repositories {
+	mavenCentral()
+}
 
-		compileOnly 'org.projectlombok:lombok'
-		annotationProcessor 'org.projectlombok:lombok'
-		testImplementation 'org.springframework.boot:spring-boot-starter-test'
-		testImplementation 'org.springframework.security:spring-security-test'
-	}
+dependencies {
+	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+	implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
+	implementation 'org.springframework.boot:spring-boot-starter-security'
+	implementation 'org.springframework.boot:spring-boot-starter-web'
+
+	/* mysql(maria DB로 변경 예정)*/
+	implementation 'mysql:mysql-connector-java:8.0.32'
+
+	/* jwt */
+	implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
+
+	compileOnly 'org.projectlombok:lombok'
+	annotationProcessor 'org.projectlombok:lombok'
+	testImplementation 'org.springframework.boot:spring-boot-starter-test'
+	testImplementation 'org.springframework.security:spring-security-test'
+}
+
+tasks.named('test') {
+	useJUnitPlatform()
 }
