--- server-api/build.gradle
@@ -1,30 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group = 'org.example'
-version = '0.0.1-SNAPSHOT'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-security'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-
-    /* jwt */
-    implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
-
-    implementation project(':server-domain')
-    implementation project(':server-global-utils')
-    implementation project(':server-infrastructure')
-
-    testImplementation platform('org.junit:junit-bom:5.9.1')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
-    testImplementation 'org.springframework.security:spring-security-test:6.0.1'
-}
-
-test {
-    useJUnitPlatform()
-}
