--- server-api/src/main/java/org/example/apis/oauth/domain/OAuth2UserInfo.java
@@ -1,16 +1,0 @@
-package org.example.apis.oauth.domain;
-
-import java.util.Map;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public abstract class OAuth2MemberInfo {
-
-    protected Map<String, Object> attributes;
-
-    public abstract String getOAuth2Id();
-    public abstract String getEmail();
-    public abstract String getName();
-}
