--- server-api/src/main/java/org/example/apis/oauth/domain/OAuth2UserInfoFactory.java
@@ -1,2 +1,0 @@
-package org.example.apis.oauth.domain;public class OAuth2MemberInfoFactory {
-}
