--- server-api/src/main/java/org/example/common/security/authentication/CustomUserDetails.java
@@ -1,2 +1,0 @@
-package org.example.common.security.authentication;public class CustomUserDetails {
-}
