--- server-api/src/main/java/org/example/common/security/filter/JwtAuthenticationFilter.java
@@ -1,22 +1,0 @@
-package org.example.common.security.filter;
-
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.ServletRequest;
-import jakarta.servlet.ServletResponse;
-import java.io.IOException;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.example.common.security.jwt.JwtProvider;
-import org.springframework.web.filter.GenericFilterBean;
-
-@Slf4j
-@RequiredArgsConstructor
-public class JwtAuthenticationFilter extends GenericFilterBean {
-    private final JwtProvider jwtProvider;
-    @Override
-    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
-            throws IOException, ServletException {
-
-    }
-}
