--- server-api/src/main/java/org/example/common/security/jwt/JwtProvider.java
@@ -1,4 +1,0 @@
-package org.example.common.security.jwt;
-
-public class JwtProvider {
-}
