--- server-api/src/main/java/org/example/common/security/oauth/handler/OAuth2AuthenticationFailureHandler.java
@@ -1,6 +1,0 @@
-package org.example.common.security.oauth.handler;
-
-import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler;
-
-public class OAuth2AuthenticationFailureHandler extends SimpleUrlAuthenticationFailureHandler {
-}
