--- server-api/src/main/java/org/example/common/security/oauth/handler/OAuth2AuthenticationSuccessHandler.java
@@ -1,6 +1,0 @@
-package org.example.common.security.oauth.handler;
-
-import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
-
-public class OAuth2AuthenticationSuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
-}
