--- server-domain/src/main/java/org/example/Main.java
@@ -1,7 +1,0 @@
-package org.example;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
