--- server-domain/src/main/java/org/example/user/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package org.example.user.repository;
-
-import java.util.Optional;
-import org.example.user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findBySocialId(String socialId);
-}
