--- server-domain/src/main/java/org/example/user/type/AuthType.java
@@ -1,8 +1,0 @@
-package org.example.user.type;
-
-import lombok.Getter;
-
-@Getter
-public enum AuthType {
-    KAKAO, GOOGLE, NAVER;
-}
