--- server-domain/src/main/java/org/example/user/type/RoleType.java
@@ -1,11 +1,0 @@
-package org.example.user.type;
-
-public enum RoleType {
-    ROLE_GUEST("비회원"), ROLE_USER("회원");
-
-    private String name;
-
-    RoleType(String name) {
-        this.name = name;
-    }
-}
