--- server-infrastructure/build.gradle
@@ -1,19 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group = 'org.example'
-version = '0.0.1-SNAPSHOT'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testImplementation platform('org.junit:junit-bom:5.9.1')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
-}
-
-test {
-    useJUnitPlatform()
-}
