+++ src/main/java/org/example/TFinderApplication.java
@@ -1,0 +1,13 @@
+package org.example;
+
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TFinderApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(TFinderApplication.class, args);
+
+    }
+}
