+++ src/main/java/org/example/domain/oauth/domain/GoogleOAuth2User.java
@@ -1,0 +1,25 @@
+package org.example.domain.oauth.domain;
+
+import java.util.Map;
+
+public class GoogleOAuth2User extends OAuth2UserInfo {
+
+    public GoogleOAuth2User(Map<String, Object> attributes) {
+        super(attributes);
+    }
+
+    @Override
+    public String getOAuth2Id() {
+        return (String) attributes.get("sub");
+    }
+
+    @Override
+    public String getEmail() {
+        return (String) attributes.get("email");
+    }
+
+    @Override
+    public String getName() {
+        return (String) attributes.get("name");
+    }
+}
