+++ src/main/java/org/example/domain/oauth/domain/KakaoOAuth2User.java
@@ -1,0 +1,28 @@
+package org.example.domain.oauth.domain;
+
+import java.util.Map;
+
+public class KakaoOAuth2User extends OAuth2UserInfo {
+
+    private final String id;
+
+    public KakaoOAuth2User(Map<String, Object> attributes) {
+        super((Map<String, Object>) attributes.get("kakao_account"));
+        this.id = (String) attributes.get("id");
+    }
+
+    @Override
+    public String getOAuth2Id() {
+        return this.id;
+    }
+
+    @Override
+    public String getEmail() {
+        return (String) attributes.get("email");
+    }
+
+    @Override
+    public String getName() {
+        return (String) ((Map<String, Object>) attributes.get("profile")).get("nickname");
+    }
+}
