+++ src/main/java/org/example/domain/user/type/AuthType.java
@@ -1,0 +1,8 @@
+package org.example.domain.user.type;
+
+import lombok.Getter;
+
+@Getter
+public enum AuthType {
+    KAKAO, GOOGLE, NAVER;
+}
