+++ src/main/java/org/example/global/security/authentication/CustomUserDetails.java
@@ -1,0 +1,83 @@
+package org.example.global.security.authentication;
+
+import java.util.Collection;
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
+import lombok.Setter;
+import org.example.domain.user.domain.User;
+import org.example.domain.user.type.RoleType;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+
+public class CustomUserDetails implements UserDetails, OAuth2User {
+
+    private Long id;
+    private String email;
+    private String socialId;
+    private Collection<? extends GrantedAuthority> authorities;
+
+    @Setter
+    private Map<String, Object> attributes;
+
+    public CustomUserDetails(Long id, String email, String socialId, Collection<? extends GrantedAuthority> authorities) {
+        this.id = id;
+        this.email = email;
+        this.socialId = socialId;
+        this.authorities = authorities;
+    }
+
+    public static CustomUserDetails create(User user, Map<String, Object> attributes) {
+        List<GrantedAuthority> authorities = Collections.singletonList(new SimpleGrantedAuthority(RoleType.ROLE_USER.name()));
+        CustomUserDetails userDetails = new CustomUserDetails(user.getId(), user.getEmail(), user.getSocialId(), authorities);
+        userDetails.setAttributes(attributes);
+        return userDetails;
+    }
+
+    @Override
+    public String getName() {
+        return socialId;
+    }
+
+    @Override
+    public Map<String, Object> getAttributes() {
+        return null;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return authorities;
+    }
+
+    @Override
+    public String getPassword() {
+        return null;
+    }
+
+    @Override
+    public String getUsername() {
+        return email;
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+}
