+++ src/main/java/org/example/global/security/jwt/JwtProvider.java
@@ -1,0 +1,7 @@
+package org.example.global.security.jwt;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class JwtProvider {
+}
