+++ src/main/java/org/example/global/security/oauth/handler/OAuth2AuthenticationFailureHandler.java
@@ -1,0 +1,8 @@
+package org.example.global.security.oauth.handler;
+
+import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler;
+import org.springframework.stereotype.Component;
+
+@Component
+public class OAuth2AuthenticationFailureHandler extends SimpleUrlAuthenticationFailureHandler {
+}
