+++ src/main/java/org/example/global/security/oauth/handler/OAuth2AuthenticationSuccessHandler.java
@@ -1,0 +1,8 @@
+package org.example.global.security.oauth.handler;
+
+import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
+import org.springframework.stereotype.Component;
+
+@Component
+public class OAuth2AuthenticationSuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
+}
